python-dev-back/python_integration/dru-app-mis_python_docker/requirements.txt
flask
gunicorn
psycopg2-binary
flasgger
flask-cors
